leo/identification/requirements.txt
PyInquirer
psycopg2-binary
pytest